requirements.txt
django==1.8.2
Pillow
django-recaptcha
